Extract Node bridge from debate agent
